refactor sidecar_interface from interface.rs to separate file
