+++ sidecar/src/interface.rs
@@ -57,42 +57,9 @@ use crate::{config, log, tracer};
 
 use crate::service::{
     InstanceId, QueueId, RequestIdentification, RequestIdentifier, RuntimeMetadata,
-    SerializedTracerHeaderTags,
+    SerializedTracerHeaderTags, SidecarInterface, SidecarInterfaceRequest,
+    SidecarInterfaceResponse,
 };
-
-#[datadog_sidecar_macros::extract_request_id]
-#[datadog_ipc_macros::impl_transfer_handles]
-#[tarpc::service]
-pub trait SidecarInterface {
-    async fn enqueue_actions(
-        instance_id: InstanceId,
-        queue_id: QueueId,
-        actions: Vec<SidecarAction>,
-    );
-    async fn register_service_and_flush_queued_actions(
-        instance_id: InstanceId,
-        queue_id: QueueId,
-        meta: RuntimeMetadata,
-        service_name: String,
-        env_name: String,
-    );
-    async fn set_session_config(session_id: String, config: SessionConfig);
-    async fn shutdown_runtime(instance_id: InstanceId);
-    async fn shutdown_session(session_id: String);
-    async fn send_trace_v04_shm(
-        instance_id: InstanceId,
-        #[SerializedHandle] handle: ShmHandle,
-        headers: SerializedTracerHeaderTags,
-    );
-    async fn send_trace_v04_bytes(
-        instance_id: InstanceId,
-        data: Vec<u8>,
-        headers: SerializedTracerHeaderTags,
-    );
-    async fn ping();
-    async fn dump() -> String;
-    async fn stats() -> String;
-}
 
 #[derive(Serialize, Deserialize)]
 pub struct SidecarStats {
