+++ sidecar/src/service/mod.rs
@@ -6,9 +6,13 @@ pub use queue_id::QueueId;
 pub use request_identification::{RequestIdentification, RequestIdentifier};
 pub use runtime_metadata::RuntimeMetadata;
 pub use serialized_tracer_header_tags::SerializedTracerHeaderTags;
+pub use sidecar_interface::{
+    SidecarInterface, SidecarInterfaceClient, SidecarInterfaceRequest, SidecarInterfaceResponse,
+};
 
 mod instance_id;
 pub mod queue_id;
 mod request_identification;
 mod runtime_metadata;
 mod serialized_tracer_header_tags;
+mod sidecar_interface;
